+++ tests/test_gsv.py
@@ -107,3 +107,23 @@ class TestGsv():
         assert data.t.mean().data == pytest.approx(279.3509), "Field values incorrect"
         data = reader.retrieve(var=130)  # test numeric argument
         assert data.t.mean().data == pytest.approx(279.3509), "Field values incorrect"
+
+    def test_reader_3d(self) -> None:
+        """Testing 3D access"""
+
+        reader = Reader(model="IFS", exp="test-fdb", source="fdb-levels", loglevel=loglevel)
+        data = reader.retrieve()
+        # coordinates read from levels key
+        assert all(data.t.coords["plev"].data == [99999., 89999., 79999.]), "Wrong coordinates from levels metadata key"
+        # can read second level
+        assert data.t.isel(plev=1).mean().values == pytest.approx(274.79095), "Field values incorrect"
+
+        data = reader.retrieve(level=[900, 800])  # Read only two levels
+        assert data.t.isel(plev=1).mean().values == pytest.approx(271.2092), "Field values incorrect"
+
+        reader = Reader(model="IFS", exp="test-fdb", source="fdb-nolevels", loglevel=loglevel)
+        data = reader.retrieve()
+        # coordinates read from levels key
+        assert all(data.t.coords["plev"].data == [100000, 90000, 80000]), "Wrong level info"
+        # can read second level
+        assert data.t.isel(plev=1).mean().values == pytest.approx(274.79095), "Field values incorrect"